Move weighted_fusion into MathTools
